Include vendor/oracle-instantclient in the LD_LIBRARY_PATH
In order for Bundler to install the ruby-oci8 gem later on it needs the
oracle-instantclient libraries in LD_LIBRARY_PATH.  It's likely that
https://github.com/heroku/heroku-buildpack-ruby would be actually
running Bundler.

Setting environment variables isn't as simple as calling `export` in
this context, because this buildpack's compile script executes in a
different sub-environment than that of the next buildpack's compile
script.  Heroku's buildpack uses the ENV_DIR mechanism to set
environment variables between buildpacks.